CommonDependencyProvider update

+++ src/main/java/cz/muni/fi/pv168/project/storage/sql/entity/RecipeIngredientAmountEntity.java
@@ -1,7 +1,7 @@
 package cz.muni.fi.pv168.project.storage.sql.entity;
 
-public record RecipeIngredientAmountEntity(Long id, String guid, long recipeId, long ingredientId, float amount) {
-    public RecipeIngredientAmountEntity(Long id, String guid, long recipeId, long ingredientId, float amount) {
+public record RecipeIngredientAmountEntity(Long id, String guid, long recipeId, long ingredientId, int amount) {
+    public RecipeIngredientAmountEntity(Long id, String guid, long recipeId, long ingredientId, int amount) {
         this.id = id;
         this.guid = guid;
         this.recipeId = recipeId;
@@ -9,7 +9,7 @@ public record RecipeIngredientAmountEntity(Long id, String guid, long recipeId, 
         this.amount = amount;
     }
 
-    public RecipeIngredientAmountEntity(String guid, long recipeId, long ingredientId, float amount) {
+    public RecipeIngredientAmountEntity(String guid, long recipeId, long ingredientId, int amount) {
         this(null, guid, recipeId, ingredientId, amount);
     }
 }
